Remove exposure-notification module. Impliment submit positive diagnosis with TemporaryExposureKey getting from ExposureNotification API.

+++ app/src/main/java/dev/keiji/cocoa/android/entity/TemporaryExposureKey.kt
@@ -13,7 +13,7 @@ data class TemporaryExposureKey(
     val reportType: Int,
 
     @SerialName("rollingStartNumber")
-    val rollingStartNumber: Long,
+    val rollingStartNumber: Int,
 
     @SerialName("rollingPeriod")
     val rollingPeriod: Int,
